price of delivery

+++ src/controller/IClientController.java
@@ -6,7 +6,7 @@ import model.Product;
 
 public interface IClientController {
 
-    int sendProductRequest(Product product, String email, Location from, Location to);
+    double sendProductRequest(Product product, String email, Location from, Location to);
 
     String whereIsMyProduct(int id);
